Make eth_chainId return uint64_t

+++ libethereum/ClientBase.h
@@ -190,7 +190,7 @@ public:
 
     Block latestBlock() const;
 
-    int chainId() const override;
+    uint64_t chainId() const override;
 
 protected:
     /// The interface that must be implemented in any class deriving this.
